DEV: Added build files and other stuff

+++ client/GUI.cpp
@@ -627,12 +627,12 @@ float GUI::getAlpha(int id)
 	return elements.at(id)->alpha;
 }
 
-float GUI::getSizeX(int id)
+int GUI::getSizeX(int id)
 {
 	return elements.at(id)->size_x;
 }
 
-float GUI::getSizeY(int id)
+int GUI::getSizeY(int id)
 {
 	return elements.at(id)->size_y;
 }
